Fix permission of comment

+++ app/controllers/comments_controller.rb
@@ -1,10 +1,13 @@
 class CommentsController < ApplicationController
+  before_filter  :authenticate_user!, except: [:index]
+
   def index
     @project = get_project
     @comments = @project.comment_threads.order('created_at desc').page(params[:page])
   end
 
   def create
+    @user = current_user
     @comment_hash = params[:comment]
     @obj = @comment_hash[:commentable_type].constantize.find(@comment_hash[:commentable_id])
     @comment = Comment.build_from(@obj, current_user.id, @comment_hash[:body])
@@ -17,13 +20,18 @@ class CommentsController < ApplicationController
 
   def destroy
     @comment = Comment.find(params[:id])
-    if @comment.destroy
-      render json: @comment, status: :ok
+    if @comment.owner?(current_user)
+      if @comment.destroy
+        render json: @comment, status: :ok
+      else
+        render js: "alert('error deleting comment');", status: 401
+      end
     else
-      render js: "alert('error deleting comment');"
+      render js: "alert('Do not have permission to delete this comment');", status: 401
     end
   end
 
+  private
   def project_id
     params.require(:project_id)
   end
